refactor: rename MainControls to ExploreControls
