+++ src/lib/controls/ExploreControls/IndicatorSelection/IndicatorParameters.svelte
